Refactor policy scanner rules into importable runtime modules
