+++ src/gabion/tooling/policy_rules/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from . import branchless_rule, defensive_fallback_rule, no_monkeypatch_rule
+
+__all__ = [
+    "branchless_rule",
+    "defensive_fallback_rule",
+    "no_monkeypatch_rule",
+]
